bring lar1ndcode up to larsoft v03_04_06. Includes fixing Geometry functions float->double, and commenting out ChanFilter

+++ lar1ndcode/Geo/ChannelMaplar1ndAlg.cxx
@@ -306,7 +306,7 @@ namespace geo{
 
 
   //----------------------------------------------------------------------------
-  float ChannelMaplar1ndAlg::WireCoordinate(float YPos, float ZPos,
+  double ChannelMaplar1ndAlg::WireCoordinate(double YPos, double ZPos,
 					  unsigned int PlaneNo,
 					  unsigned int TPCNo,
 					  unsigned int cstat) const
